Fix incorrect values in cableConnected event.
Explanation from Franky why is this needed and why we cant just
parse uevent data (and verified by me - radekp):

You can't control the sequence uevents are coming in, so for a usb cable
insertion (connecting to the pc) you have at least one event for usb and one
for ac. One says "1", the other says "0" ... so you can end up with the phone
thinking it is still not connected. What I do is in fact checking the
cablestatus before setting cableconnected to 1 (or 0). This is flawless, works
on all kernels (even buggy ones, I've tried several this weekend) and btw: the
same logic is used in neobattery as well :-)