When kbd handler searches in /dev/input close not matching files
The neokbdhandler left open files in /dev/input. The two files are
inputs from accelerometers with lot of input data, so this was
slowing down Qtopia.

One should always check if Qtopia does not leave open files that
it does not need. It's quite simple. Find pid of qpe and do ls -la
in /proc/qtopia_pid/fd