QMplayer - fix fullscreen flickering
When user switched to fullscreen the screen started to flicker and
it has never stopped to flicker even after leaving fullscreen.

This was caused with incorrect fb mode. We now use the default mode
for full vga (output of fbset without args) and mode with increased
timings to 100000 for qvga to fix bright colors.