Patch for no incomming SMS workaround/debugging
Dirty patch that should help with no incomming SMS issue. It writes
/tmp/newMessageArrived.txt when modem notices +CMTI and small label
is displayed on home screen, if Qtopia does not notice new SMS
message.

If this works then we know that hardware and AT handling is OK and
problem must be somewhere after noticing +CMTI and displaying SMS
message dialog.