Remove setting sd_max_clk because it makes some SD cards unusable.
Here is mail from Alex about it:

MHO its not very good idea. My card (Kingston 8GB SHDC) works bad with
this value, and it was not very easy for me to find why its changing on
every restart :) I think that anyone who need this able to change value
in sysctl, or kernel boot strint, or init.d.

Also i found that if i`m changing value in shell (after boot) then it
affects only after suspend/resume cycle. Thats another reason why i
decided to keep it in kernel boot args.

And also reply from Vencabot Teppoo:

This solution worked for me. I'm genuinely surprised, but I doubled the
max click as suggested in this post and I made the EXT3 fs as usual,
but, when I checked it, it was clean. I mounted it, saved a file on it,
unmounted it, remounted it, and the file was still there.

Thanks Vencabot for reporting this.