Reconnect wifi after resume.
Provided fix is done to solve issue with WIFI and Sleep.

Whats going on - when system go to the suspend state eth0 device is
removing. Then its adding again on resume. udhcp goes crazy because of
this, and not only it. Static ip configuration is also broken. Another
problem that its not checking if connection is alive after wakeup.

What my patch do:

1) Using udev rule it run special script on resume.
2) If connection was active on suspend time it stops connection and trying
to open it again. System trying to reconnect (using qcop), and if its not
possible - changes status to offline.

It works for me. Sometime it does not reconnect, but this is probably
because of AR6000 bugs. At least it correctly goes to the offline state
then.