Fix voice calls during GPRS session.
I have another GPRS issue - when data session is active there is no
voice calls are received.  I was thinking that it is modem bug, but on
chat debug log i found that modem actually ringing. More then this -
after such attempt GPRS goes offline and to put it online qtmoko restart
is required. Below is an example:

1) We are going to GPRS/online.
2) Attemting to call to the device. Device is not ringing, remote
getting long calls. After hangup from remote - gprs icon goes off (pppd
is still on!).
3) In the internet menu its not possible to turn GPRS on - it simply
ignores status change.  Need to reboot qpe to make it working again.

 From ATCHAT i see that device was ringing, so problem is not with
multiplexor or modem. I think its some kind of logical bug in the class
responsible for call processing, but it needs more investigation, i`m
not sure whats going wrong. Anyway - good news that its possible to
receive voice calls with GPRS enabled, and bad news that it fail in
qtmoko :)

snip ...

After long investigation i found the problem source. In the
/devices/neo/src/plugins/phonevendors/neo/vendor_neo.cpp there is a
function to receive the calls (NeoCallProvider::cpiNotification). It
puts identifier to the

ringing( QAtUtils::decodeNumber( number, type ), callType, identifier );

function, and also check it in the

QModemCall *call = callForIdentifier( identifier );

And this is a problem. GRPS session also has identifier and it is 1 (the
same we are getting in the %CPI response. I was not able to find where
netmanager sets it, but i`m sure that it is 1, because if gprs enabled
QModemCall *call = callForIdentifier( identifier ); is true (call
active), so it never calling announceCall() and call goes to nowhere.
More then this - it later (after ringing) trying to hangup gprs session,
what causing GPRS bug.

I`m not sure that my fix is 100% correct (because i have no idea how
this should work and where gprs session identifier is set), but at least
it works fine with a few tests i done. I just replaced

uint identifier = QAtUtils::parseNumber( msg, posn );

with

uint identifier = QAtUtils::parseNumber( msg, posn )+1000;

And it starts working . As for me - its better to redefine gprs call
id,  but i was not able to find where it is.

Some tests i done:

1) No GPRS: voice working correctly
2) GPRS enabled and idle - phone starts ringing, i can take a call,
speak. In the time of the call gprs is not working, but after hangup -
starts working again (modem limitation)
3) If i have active download then no incoming call is going to the
device - it is also modem limitation, i see from atchat log that there
is no RINGING.
4) After all this tests its possible to stop and start gprs again, no
problems found.

I think that this ID issue was also cause of other problems related to
GPRS.