Faenqomod theme
Hi list!
This is an update of the faenqomod theme. It is based on the wonderful
faenqo theme by cyberesprit and presents some variations.
As we are collaborating, many of the features of this release of faenqomod
could be included in the future releases of the original faenqo (and vice
versa). The main project is hosted by cyberesprit, which I thank, on
https://cyberesprit.fr/faenqo/
In this release of faenqomod you will see:
- a different homescreen layout then the original faenqo;
- different animation for battery in charge and in discharge;
- new context buttons;
- almost EVERY icons substituted with the Faenza icons
(http://tiheum.deviantart.com/art/Faenza-Icons-173323228) and icons inspired
on them;

For now you'll find faenqomod, screenshots, and faenqo svg icons at
http://db.tt/vG5D0yZ
To install faenqomod it is necessary to copy/paste the files provided in the
archive. I'll try to make a deb in the next days... but if someone is faster
then me, he is welcome :)
enjoy!

Joif