Fix "No network" bug
If network registration is already in registered state before
CellModemManager is created manager will never realize that it
is already registers to network because the network state change
event is never called.

This should fix it - i tried just with FSO phone provider but it
could also fix it for AT based modems.