FSO - discard status results if event ocurred during the async call
This avoid races cause with async status calls. E.g. when we initialize
we call GetDeviceStatus(). The call is now running on background and
will return later e.g. "uninitialized".

The call is still in progress while DeviceStatus signal is emited with
new status e.g. "initialized". We process this new state.

After that GetDeviceStatus() finishes with still old and no longer
valid status "uninitialized".

We fix this with bool variables which will discard GetDeviceStatus()
call if event ocurred during the call.

Other solution would be to call GetDeviceStatus() synchronously which
should be ok and fast according to morphis on IRC.