qtmoko: Enlarge qtmail UI folders
This tiny patch enlarge the folders in qtmail, so those
are bigger and easier to press. With the new size the
list of folders uses the whole screen (there was a useless
free space at bottom).

Before this patch the folders at that ui looks like this :
http://fz.hobby-site.org/hp660lx/qtmoko/previous_size_of_qtmail_folders.png
After this patch the ui looks like this :
http://fz.hobby-site.org/hp660lx/qtmoko/new_size_of_qtmail_folders.png