qtmoko: autorotation: Add the autorotation program which rotates the whole QT GUI
This serie of patches add the autorotation feature to the qtmoko project.

The autorotation features is :

- this patch with the autorotation program
- the next patch with the autorotation perl script and
  two scripts to turn on and off autorotation feature

The autorotation feature works like this :

The main script sleeps most of time. After 3 seconds sleeping it reads
accelerometers, doing 5 readings.

If the phone was rotated by the user it calls autorotation program
and rotates the whole QT GUI. Then the main script
starts to sleep again for 3 seconds. If the phone was
not rotated by the user, then it continues sleeping.

The idea is to be quiet and dont consume any cpu, or just
a bit every time the user rotates the phone.

To turn on or off the autorotation feature the user will find
the options ON and OFF in the services list (where you
find wifi on/off, gps on/off, etc).

A demo of this feature :

http://www.youtube.com/watch?v=UrcMzIAM8N4