applicationlauncher - make second launching of program with args working
Use exe + params as identifier for running process. Before it was
mismatched - for registering we use exe+params and for state changes
used only exe.

So we unify this to exe+prarams so that it allows us to run two exes
with different params at the same time.

e.g.

    qterminal -c script1.sh && qterminal -c script2.sh

should now work.