Mokofaen theme
Hi people
There is a new theme that I worked on: MokoFaen.
It is the continuation of the work done on Faenqo/Faenqomod and it is
inspired, in part, from the today smartphone GUIs (just to not say that...
it shamelessly copies them ;) ).
This theme is tested on QtMoko v35. I will flash the v37 on my Neo soon, I
will submit updates if necessary for this version.
Major changes from Faenqomod:
- some icons updated from the recent release of Faenza icons set;
- introduced new icons where needed;
- a completely new Home theme [1];
- Dialer screen revised (bigger buttons but without the call-history and sms
icons in the screen);

You'll find the files/source (and some screenshots) here [2].
To install the theme extract MokoFaen.tar.gz and copy/merge the directory on
your moko.
Waiting for an enough brave man who can package it and upload it in qtmoko
apps page (:
As it is a sort of Faenqomod v2 I suppose that Faenqomod can be possibly
deleted from the apps page.

There is a question: some parts of the theme are taken from other works,
IINW released under GPL2/3. Is it sufficient to attach the GPL license in
the source?

Best regards
Joif

[1] you can return to the "classic view" simply by replacing (in
/opt/qtmoko/etc/themes/mokofaen/) home.xml with home_classic.xml. Also you
can disable the clock by commenting out its lines in home.xml.
It could be cute to have a proper option to do these in the "appearance"
app.
[2] http://db.tt/8KF7DHRp