Stop forcing Qt rebuild everytime qbuild is run (bug #45).
This was the reason of zillions wasted CPU cycles, and endless frustration.
The direct consequence will be that if you do changes to Qt itself
(in qtopiacore/qt), qbuild will not see them; you will have to rerun configure,
and then qbuild will propagate you Qt changes to QtExtended stuff, unleashing
the wrath of a full rebuild (but only once, and you asked for it ;).