Let nct_ and theme_lupdate pass "-locations none" to lupdate.
Since those runs are done on generated files, the location info is strictly
non-informational and only brings in loads of noise.

A long-term fix would be to add #line" directives in the generated cpp files,
and let lupdate use those, but it's not necessary for a first step.

Another option would be to rewrite those lupdate wrappers as directly
producing the XML, instead of being wrappers.  Since lupdate is much more
monolithic than gettext, it would require to link to lupdate code and would
probably more work overall, including long-term maintainance :(