Avoid getting blocked waiting for too long in eventloop (issue #69).
For some reason processEvents does not return at once once the neo modem
is suspended.

While the WaitForMoreEvents is probably required to avoid running in a tight
loop while the handlers get run, setting a 0.1s timeout prevents waiting for
too long for any events that are of no concern to us, and give the neo instant
suspend/resume.

But it still looks like a workaround, why would processEvents wait for more
events after it has handled one event, that's not the documented semantics for
WaitForMoreEvents.