gta04 gps plugin - make sure the gps is always turned on
After start we check after 3s if we have data on serial port. If not
call the poweron script again, because the poweron works actually as
toggle.

This fixes situation where we start and GPS is already turned on. In
this case ctor call to poweron turns if actually off, so we call
poweron again.