Fixed german svg keyboard layouts
Hi Radek,

I tried the de-layouts of the svg-Keyboards that are available at
http://qtmoko.sourceforge.net/apps/qtmoko-kbd-de.html
They look close to a german layout, but they still return the same keys
as an american layout would. E. g. when I typed "Ö" I got ":" So I fixed
that in the svg files (see attachment).

Regards
Peter