whereabouts - use process for reading NMEA from serial port
Since Qt 4.7 the condition !d->m_source->atEnd() in
QNmeaWhereabouts::newDataAvailable stopped working.

We are using QFile for reading from the serial port. It does not emit
read ready event and it's difficult to read all available data (when
you get to the end it always block and this block whole UI).

Simple solution is to use external process for reading from the port.
It emits read ready signal and readLine() returns 0 when all data is
read.

This fixes GPS on GTA04.

I have also implemented notification on POWER led.