alarmcontrol - fix and cleanup alarms
Alarms were broken. They rely on bundled atd crap which is used just
to wake the phone up from suspend. When phone is woken the code checks
for nearest alarm and starts ringing if found.

The problem is that wakeup interrupt triggered by atd is using RTC time
and the code for alarm check is using system time and Qt timers. If RTC
and system time get out of sync the phone is put in suspend and never
wakes up by RTC anymore.

So the fix is to also remember the alarm time in RTC clock time and check
for both. We also do the check for alarm after resume.

Another nice thing now is that we can get rid of bundled atd which was
holding /dev/rtc0 open and utilities like hwclock were not working because
of this.