gta04 - read and report via led interrupts from modem
We are reading /dev/input/incoming which generates input event on
incoming call or SMS.

When new event arrives we set fast blinking on the AUX green led.
If it apppears to be call then we switch to slow blinking (already
implemented in qmodem after incoming call AT command arrives).

If it's not call we leave it fast blinking and this indicates SMS.