gta04 - Don't wait for response to AT+VTS command (for DTMF tones)
Currently, within a given call, only the first DTMF tone goes out,
I believe because the GTA04 modem doesn't send any response to AT+VTS,
and so the AT chat gets stuck.  The patch below fixes that by adding a
parameter to QAtChat::chat that tells it not to wait for a response.

Without this change, it's only possible to send one DTMF tone in a
call - because after that the AtChat is stuck waiting for a response
to the AT+VTS command for that first one.  With this change, I've
tested that it's possible to send multiple tones within the same call,
and that they have the expected effect (as determined by the remote).