qt - Enable support for Windows charsets
Without this change, emails (or email parts) with

Content-Type: ...; charset=WINDOWS-...

are displayed as though they're empty.  In Qt, charsets are known as
codecs, and the support for WINDOWS-... charsets is in a file named
qsimplecodec.cpp, but only compiled in if QT_NO_CODECS is not defined.

Therefore, to get support for WINDOWS-... charsets, we modify
qconfig-qpe.h so as not to define QT_NO_CODECS.