QtMaze: use accelerometer callback instead of running own timer
I have one more accelerometer-patch, below.  It changes QtMaze so that
it doesn't run its own fast timer in parallel with the one in the
accelerometer library.

I've had this for a while, so I'm confident it's good for GTA04.  It
would be good to hear from GTA01/2 people that all accelerometer
functions are still working correctly there.