For A3, set 'AVADC Clock Priority' control to 'HiFi high priority'
Per advice from Andreas Kemnade:

> > Very important should also be the clock priority setting
> > (AVADC Clock Priority).
>
> Can you say more about that?  All the QtMoko state files have value
> "Voice high priority".  Is that correct?
>
It is about the samplerate of the ad converter in the twl4030. It gets
the samplerate of the higher priority path and then it gets resampled in
poor quality for the other path.  If I record audio on the
gta04 and play it back I hear a clear difference. If you use the
audio path to record data, it should be "HIFI high priority".
If you use the voice path (hardware routing on the GTA04A4,
"Voice high priority" is right).