Dont use pkg-config, specify all lib and include dirs in  mkspecs
pkg-config with QtMoko is really not working. The PKG_CONFIG_PREFIX
and it's handling in configure looks like hack. In the end i was
always getting some crappy -L and -I dirs in gcc args.

We might have need multiple mkspecs for each debian version, but
it will be more straightforward and robust.