Use Debian's libvorbisfile library instead of QtMoko's bundled and old 'tremor'
This integrates better into Debian and reduces QtMoko's size.  It also
fixes segfaults (following glibc double free or corruption reports)
that are seen when scanning .ogg files since QtMoko switched to using
gstreamer as its media engine.

My guess is that the segfaults were caused by some subtle
incompatibility between the integer-based (i.e. tremor/libvorbisidec)
and the float-based (i.e. libvorbis and libvorbisfile) versions of the
Ogg Vorbis libraries.  For example, tremor/libvorbisidec and
libvorbisfile provide incompatible versions of the OggVorbis_File
type.  The cruxus media engine and oggcontentplugin (as first coded)
both used the tremor library in QtMoko's codebase, and that all
appeared to work fine.  The introduction of gstreamer and its
Ogg-support plugin pulled in the float-based Ogg libraries, and since
then (I believe) we've been seeing faults from oggcontentplugin.
Switching oggcontentplugin to use the float-based
libraries (libvorbisfile) appears to fix those faults.

I have not been able to pin down exact details of this problem.
Specifically, apart from the new build of oggcontentplugin, there are
no QtMoko executables or shared objects that link to libvorbisfile; so
I don't think we could ever have been using a float-based
implementation of the OggVorbis_File type with an integer-based header
for it.  However, it seems plausible that there are lower-level
functions or types deeper inside libogg+libvorbis and libvorbisidec,
with the same names but incompatible integer-based and float-base
implementations, and those could account for the problems that we've
observed.