fix "Draft message" problem
With Qt 4.8 there is this regression:

When messages are retrieved from SIM they have in "From" field text
"Draft message" instead of sender number . You can quite easily
reproduce it by stopping qtmoko and running:

    rm -rf /home/root/Applications/qtopiamail/

The problem is that incoming flag is lost during SQL update of the
"mailmessages" table and this is caused by upgrading sqlite to newer
version. Qt 4.7 uses older sqlite then Qt 4.8.

The interesting function is QMailStore::attemptUpdateMessagesStatus()

It executes following SQL:

    UPDATE mailmessages SET status = status & ~status_flag WHERE

status argument is in C++ quint64 but in the database it's INTEGER.

Sqlite has variable size integer and the ~ operator behaviour is not
working well in the new version - and also it's not very correct. It's
much safer to use UNSIGNED BIG INT so that C++ and SQL match.