Really fix "Draft Message" problem
Following Radek's identification of the problem being how sqlite
handles ~status, where status is a quint64, this change masks off the
bits of ~status that have no meaning, leaving a much smaller number
for sqlite to deal with.  For me this really appears to solve the
Draft Message problem.

For original problem see:

https://github.com/radekp/qtmoko/commit/b731ab616d486649f0b9c3fba9b32604a135849a