gps - use qgpsdwhereabouts everywhere
This fixes navit slow exit and crash. For more info:

http://lists.goldelico.com/pipermail/gta04-owner/2013-March/004155.html

I have decided to not use gpsd hooks because they are not available
in sqeeze and i also dont like the idea of setuid rfkill. It will be
no problem for QtMoko apps and you can easily call gps-poweron/off
manually if needed.