Optimise 'Removed' flagging of messages deleted from the IMAP server
On each IMAP server sync, if there are messages that exist on the
phone but have since been deleted from the IMAP server, or moved to a
different folder on the IMAP server, QtMoko adds the
QMailMessage::Removed flag to those messages.  Note that we don't
automatically delete those messages from the phone.

I once - not using QtMoko - moved several thousand older messages to
an "OLD" folder on my IMAP server.  After that I found that each
QtMoko IMAP server sync operation would take a lot of time and CPU,
and discovered this was because it was reflagging all of those moved
messages every time.  If we optimize the process by skipping messages
that already have the 'Removed' flag, it goes massively faster.