fix to get rid of 'WARNING: /qtmoko/src/build/qt_patch/default_post.prf:115: Unescaped backslashes are deprecated.'
Ther are some places where a backslash is used as a Windows style
file path. Since we neither build on Windows nor have a problem
we can convert to standard slashes.

The line 115 error is because the code tries to find out
how to initialize a variable with CRLF. It first tries to use
the backslash notation and if that fails it uses a function to
convert a string with explicit backslashes.

Both cases result in the Windows backslash warning.

Signed-off-by: H. Nikolaus Schaller <hns@goldelico.com>